+++ lib/src/entity.dart
@@ -36,7 +36,7 @@ class AssetPathEntity {
   /// path asset type.
   RequestType _type;
 
-  final FilterOptionGroup filterOption;
+  FilterOptionGroup filterOption;
 
   /// The value used internally by the user.
   /// Used to indicate the value that should be available inside the path.
@@ -66,24 +66,34 @@ class AssetPathEntity {
 
   /// Call this method to update the property
   Future<void> refreshPathProperties({
-    DateTimeCond dateTimeCond,
-    DateTimeCond updateTimeCond,
+    FilterOptionGroup filterOptionGroup,
+    bool maxDateTimeToNow = true,
   }) async {
-    dateTimeCond ??=
-        this.filterOption.createTimeCond.copyWith(max: DateTime.now());
-    updateTimeCond ??=
-        this.filterOption.updateTimeCond.copyWith(max: DateTime.now());
+    assert(maxDateTimeToNow != null);
+
+    filterOptionGroup ??= this.filterOption.copyWith();
+
+    if (maxDateTimeToNow) {
+      filterOptionGroup = filterOptionGroup.copyWith(
+        createTimeCond: filterOptionGroup.createTimeCond.copyWith(
+          max: DateTime.now(),
+        ),
+        updateTimeCond: filterOptionGroup.updateTimeCond.copyWith(
+          max: DateTime.now(),
+        ),
+      );
+    }
 
     final result = await PhotoManager.fetchPathProperties(
       entity: this,
-      dateTimeCond: dateTimeCond,
-      updateTimeCond: updateTimeCond,
+      filterOptionGroup: filterOptionGroup,
     );
     if (result != null) {
       this.assetCount = result.assetCount;
       this.name = result.name;
-      this.filterOption.createTimeCond = dateTimeCond;
       this.isAll = result.isAll;
+
+      this.filterOption = filterOptionGroup;
     }
   }