+++ CHANGELOG.md
@@ -55,6 +55,7 @@
   - Support clean file cache.
   - Experimental
     - Preload image (Use `PhotoCachingManager` api.)
+  - Add `OrderOption` as sort condition.
 
 - Fixes
   - [#362](https://github.com/CaiJingLong/flutter_photo_manager/issues/362)